[NEW] Rewrite alignment-translate -- handle reverse-complement reading frames
This commit modifies the Codon alphabet so that it is can be used for
translating 3-tuples of nucleotides instead of just translating Codons.
This allows us to load the input data as nucleotides, and so we can be more
intelligent about handling different frames and characters that hang off the end.

More importantly, it allows us to handline reading frames -1,-2, and -3, which
require taking the reverse complement of the DNA sequence.

I have also modified alignment-translate to default to an amino-acid alphabet
with stop codons, so that these can be handled by default -- this seems the
most useful option.

* tools/alignment-translate.C
  + parse_cmd_line( )
     - "alphabet": remove option.
     - "genetic-code": new option.
     - "reverse-complement": new option.
  + main( )
    - rewrite.

* alignment-util.{H,C}
  + alignment reverse_complement(const alignment& ): new function.
  + alignment complement(const alignment& ): new function.
  + alignment reverse(const alignment& ): new function.

* alphabet.{H,C}
  + Nucleotides::complement( ): new member function.
  + vector<vector<vector<int>>> Triplets::codon_table: new data member.
  + Triplets::get_triplet(int,int,int): new member function.