Remove internal node sequences from partitions w/o an indel model
This requires extending load_As_and_T( ) and load_As_and_random_T( )
to allow a form that takes a vector<bool> that indicates whether or
not each alignment should have internal node sequences.

It also requires parsing the indel model *before* loading the
alignments in bali-phy.C

The rest of the patch simply cleans up assert( ) statement that relied
on always have a sequence in the alignment for every tree node.