Improve choose_MH( ) to get higher acceptance rates.
The old version would have a 50% chance of accepting {1}->{1,0} whereas
this version has a 100% chance.

This is a complete rewrite of choose_MH( ) to use a different algorithm.
It is based on decomposing a variable Y = [0,N-1) where P(Y_k) is
monotonically increasing into a collection of binary variables
X_k = I(Y > k).  To propose a change from i, we iteratively propose to flip
X_k for k<=i.  If some proposal is accepted, we stop there. If not, then we
Gibb's sample from the values of Y that are more than i.

This proposal always results in a different value of Y unless the
current value of Y is the most probably value.

When Y can take on only 2 values, it reduces to a normal MH move.

This algorithm has the property that Pr(i->j)/Pr(j->i) = P[j]/P[i]
and so the details are unimportant in many situations.