Make PSRF-80% usable for discrete variables ignoring differences of <= 1.
The PSRF is basically the width of the 80% confidence interval for the
union of the data sets divided by the average 80% confidence interval
over the data sets.  This measure doesn't really work for distributions
on integers because the distributions that are arbitrarily close may
still have 80% CI intervals that differ by 1.  Then the PSRF comes out to
be about 1.0 + 1/width, and may not near 1.0.

Therefore we bias this estimate downward for distributions on integers
by subtracting one from the width for the combined distribution,
in order to get a lower bound.

I suppose we could report an interval between [a,b] where

a = max(0, (total_CI-1)/ave_CI)
b = total_CI/ave_CI.

However, the current approach does what we want, which is to require
integer variables to be different by greater than 1 before reporting
trouble.

(For the integer variables, we compare the probabilies in the
intervals, which do not vary so discontinuously.  See
PSRF-RCF in statreport output.)