Increase the LC length lazily, remove explicitly length setting.
Since we currently increase the size of the cache monotonically,
we can simply ask for any extra space we might need right before
using it.

There are a small number of places that use it, all called from
calculate_caches( ).

This change is also necessary for correctness (now) since we
weren't increasing the LC length in sample-node.C and
sample-two-nodes.C.