Remove primordial, problematic code from sample_alignments();
I think this code was there original to avoid handling zero-length
alignments, since the sequence-length probabilities were originally
incorrect for zero-length sequences.  (In the RS05 model).

However, the function lengthp( ) is correct for the RS07 model,
and probably also correct for the RS05 model.  Hence killing this code
should be correct.