Disable the move if there are any fixed-alignment partitions.
In fixed-alignment partitions where the alignmnt is represented in terms
of pairwise alignments (i.e. not SEV) we need to know the path from the
source branch to the target branch in order to fix up pairwise
alignment.

But at the moment we don't have that.  I think the sample_SPR_nodes( )
move constructs it internally, though.