Not all regs with no forces started with a positive force count.
Regs that were executed _between_ the two programs might have started
with a force count of 0.  So we need to scan those regs for 0s after
incrementing counts. Possibly such regs will correspond to heads that
were executed in intermediate tokens.

Lots of regs with an eventual force count of 0 will have positive force
count after incrementing counts.  So, we put need to iteratively
decrement their counts again as we move the steps out of the token.